refactor: rename file to match function name
Relates #166
