+++ utils/messaging/insertMessage.ts
@@ -1,4 +1,4 @@
-import newClient from '../../config/supabaseclient';
+import newClient from '../../config/createClient';
 
 export default async function insertMessage(
   userId: string | undefined,
@@ -18,7 +18,7 @@ export default async function insertMessage(
 
     if (messageError)
       throw new Error(`Error inserting message: ${messageError.message}`);
-    
+
     const { error: updateError } = await supabase
       .from('user_conversations')
       .update({ has_unread_messages: true })
