+++ utils/supabase/deleteItems.ts
@@ -1,4 +1,4 @@
-import newClient from '@/config/supabaseclient';
+import newClient from '@/config/createClient';
 
 // note for if we are turning row level security back on and this breaks If you use delete() with filters and you have RLS enabled, only rows visible through SELECT policies are deleted. Note that by default no rows are visible, so you need at least one SELECT/ALL policy that makes the rows visible.
 export default async function deleteItems(columnCriteria: number) {
